refactor: page type 중복 코드 제거
