+++ src/apis/consumer/consumerAPIservice.types.ts
@@ -1,3 +1,4 @@
+import { Page } from "@/constants/PageResponseType";
 import { POINT } from "@/constants/PointEnum";
 
 interface ApiResponse<T> {
@@ -6,35 +7,6 @@ interface ApiResponse<T> {
   detail?: string;
   data: T;
   failure?: string;
-}
-
-export interface Page<T> {
-  content: T;
-  pageable: {
-    sort: {
-      empty: boolean;
-      sorted: boolean;
-      unsorted: boolean;
-    };
-    offset: number;
-    pageSize: number;
-    pageNumber: number;
-    paged: boolean;
-    unpaged: boolean;
-  };
-  last: boolean;
-  totalElements: number;
-  totalPages: number;
-  size: number;
-  number: number;
-  sort: {
-    empty: boolean;
-    sorted: boolean;
-    unsorted: boolean;
-  };
-  first: boolean;
-  numberOfElements: number;
-  empty: boolean;
 }
 
 export type GetMyInfoResponseData = {
