+++ src/apis/seller/sellerAPIService.types.ts
@@ -14,4 +14,15 @@ export type GetSellerInfoResponseData = {
   storeImageUrl: string;
 };
 
+export type GetSellerListResponseData = {
+  sellerId: number;
+  email: string;
+  businessmanName: string;
+  storeName: string;
+  storePhoneNumber: string;
+  createdAt: string;
+  storeDescription: string;
+  storeImageUrl: string;
+};
+
 export type GetSellerInfoResponse = ApiResponse<GetSellerInfoResponseData>;
